+++ android/src/main/java/io/parity/signer/screens/keysets/create/backupstepscreens/NewKeysetSelectNetworkScreen.kt
@@ -56,7 +56,7 @@ fun NewKeySetSelectNetworkScreen(
 	navigator: Navigator,
 	onBack: Callback,
 ) {
-	val networksViewModel: NewKeySetNetworksViewModel = viewModel()
+	val networksViewModel: NewKeySetNetworksWithNavigatorViewModel = viewModel()
 	val selected: MutableState<Set<String>> =
 		remember {
 			mutableStateOf(
