+++ android/src/main/java/io/parity/signer/screens/scan/bananasplit/BananaSplitSubgraph.kt
@@ -14,7 +14,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import io.parity.signer.domain.Callback
-import io.parity.signer.screens.keysets.restore.RecoverKeysetSelectNetworkScreen
+import io.parity.signer.screens.keysets.restore.recoverkeysetnetworks.RecoverKeysetSelectNetworkScreen
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
